requirements.txt
requests==2.31.0
selenium==4.34.2
schedule==1.2.1
Pillow>=10.0.0
webdrivermanager_cn==2.4.0
webdriver-manager==4.0.2
# onnxruntime==1.18.1  # 已移除，使用NCC算法替代
numpy==1.26.2
# python-dotenv
# python-dateutil